# perf: object hash