--- a/index.js
+++ b/index.js
@@ -104,13 +104,16 @@ function applyDefaults(object, sourceOptions){
   return options;
 }
 
+var nativeFunc = '[native code] }';
+var nativeFuncLength = nativeFunc.length;
+
 /** Check if the given function is a native function */
 function isNativeFunction(f) {
   if ((typeof f) !== 'function') {
     return false;
   }
-  var exp = /^function\s+\w*\s*\(\s*\)\s*{\s+\[native code\]\s+}$/i;
-  return exp.exec(Function.prototype.toString.call(f)) != null;
+
+  return Function.prototype.toString.call(f).slice(-nativeFuncLength) === nativeFunc;
 }
 
 function hash(object, options) {